Fixed build errors

+++ src/diagnostics_component/diagnostics_workflow/src/diagnostics_workflow.c
@@ -108,7 +108,7 @@ done:
  * @returns a value of Diagnostics_Result indicating the status of this component's upload
  */
 Diagnostics_Result DiagnosticsWorkflow_GetFilesForComponent(
-    VECTOR_HANDLE* fileNames, const DiagnosticsLogComponent* logComponent, const unsigned int maxUploadSize)
+    VECTOR_HANDLE* fileNames, const DiagnosticsLogComponent* logComponent, const long long maxUploadSize)
 {
     if (logComponent == NULL || maxUploadSize == 0 || fileNames == 0)
     {
@@ -311,7 +311,7 @@ void DiagnosticsWorkflow_DiscoverAndUploadLogs(const DiagnosticsWorkflowData* wo
         goto done;
     }
 
-    const unsigned int uploadSizePerComponent = workflowData->maxBytesToUploadPerLogPath;
+    const long long uploadSizePerComponent = workflowData->maxBytesToUploadPerLogPath;
 
     if (uploadSizePerComponent == 0)
     {
