+++ src/diagnostics_component/utils/file_info_utils/src/file_info_utils.c
@@ -194,7 +194,7 @@ done:
  * @returns true on successful scanning and populating of filePathVectorHandle; false on failure
  */
 bool FileInfoUtils_GetNewestFilesInDirUnderSize(
-    VECTOR_HANDLE* fileNameVector, const char* directoryPath, const unsigned int maxFileSize)
+    VECTOR_HANDLE* fileNameVector, const char* directoryPath, const long long maxFileSize)
 {
     bool succeeded = false;
